wolf.commit

+++ src/com/company/Meal.java
@@ -1,5 +1,6 @@
 package com.company;
 
 public interface Meal {
-    double getWeight(); //в интерфейске все методы с модификатором доступа public.
+
+    int getWeight(); //в интерфейске все методы с модификатором доступа public.
 }
